Refactor: 스터디 일정 코드 리팩토링

+++ MSA/Schedule/src/main/java/com/springcooler/sgma/studyscheduleparticipant/command/domain/aggregate/StudyScheduleParticipant.java
@@ -29,7 +29,7 @@ public class StudyScheduleParticipant {
     private Integer numSubmittedProblems;
 
     @Column(name = "TEST_SCORE")
-    private Double testScore;
+    private Integer testScore;
 
     @Column(name = "TEST_PERCENTAGE")
     private Double testPercentage;
